refactor(core): split audio unit tests by concern
Move the 2.2k-line audio/tests.rs dump into audio/tests/{resample,pcm,
decode,telephony,opus,stream}.rs. Shared WAV helpers stay in tests/mod.rs.
